Test/pretestnet verification 2026 07 22 (#135)
* fix(ci): restore clippy gate on main

* test: align compressed entity fixtures with schema v3

* test: restore consensus dependency boundary audit

* fix(tokens): restore mock USDT deployment guard

* fix(deps): update vulnerable runtime dependencies

* fix(consensus): avoid healthy DKG share disclosure

* fix(node): cancel stale compressed-tree payload jobs

* fix(compressed-entities): preserve stale parent readiness

* fix(evm): preserve typed hook readiness errors

* fix(node): retry stale payload attempts safely

* fix(localnet): persist TEE state across full restart

* fix(node): build payloads on canonical ancestors

* test(e2e): wait for pending tribute offers

* style: format regression fixes

* fix(consensus): advance late verifier rotations

* test(e2e): exercise late warm promotion

* docs(adr): reconcile workspace coverage ledger

* test(fuzz): make participation target runnable

* fix(cli): reject malformed private key lengths

* test(e2e): keep governance votes inside window

* test(e2e): require lifecycle tribute execution

* fix(ci): restore nightly udeps compatibility

* fix(ci): avoid future-incompatible bail expressions

* fix(ci): update offchain reader atomic API

* fix(ci): make error branches nightly compatible

* fix(ci): modernize engine error expressions

* fix(ci): modernize node backend error

* fix(contracts): refresh intent ABI exports

* fix(ci): make prerelease container build reproducible

* fix(ci): retain precompile interfaces in container context

* fix(ci): include libc++ in container runtime

* perf(ci): keep container recipe out of source layer

* docs(test): record pretestnet verification verdict

+++ crates/system/update/src/runtime.rs
@@ -64,7 +64,7 @@ impl Update<'_> {
             return Err(UpdateError::ActivationConflict.into());
         }
 
-        let waiting_len = self.waiting_for_activation_proposal_ids.len()? as u32;
+        let waiting_len = self.waiting_for_activation_proposal_ids.len()?;
         if waiting_len >= MAX_WAITING_FOR_ACTIVATION_UPDATES {
             return Err(UpdateError::TooManyWaitingForActivation.into());
         }
